attendance_system/requirements.txt
django
opencv-python
numpy
tensorflow
pandas
pillow
opencv-python-headless
djangorestframework 
keras-facenet 
mtcnn
scipy